// FinalProject/TrafficOffenceManagementSystem/TrafficOffenceManagementSystemFrontend/src/app/model/fine.ts
export class Fine {

    fineId: number;
    licenseNo: string;
    nic: string;
    vehicleNo: string;
    dateAndTimeOfOffence: Date;
    place: string;
    offenceId: number;
    validFrom: Date;
    validTo: Date;
    court: string;
    courtDate: Date;
    policeStation: string;
    issuingOfficer: number;
    status: string;
}